see extended description
- fixed bug: "Cannot exit from MainScreen by pressing the Back button if
the screen orientation (at the moment the Back button is pressed)
differs from the screen orientation when application was started.

- PartnerPointsInfoPanelFragment doesn't show the button to call phone
if the displayed partner point has not phones and it shows 1 button
otherwise (it shows single button even if the displayed partner point
has several phones, in this instance it shows
MultiPhonesCallerDialogFragment by clicking on the button to call
phone).

- refactoring